Made the error you get when a tag doesn't expand right more verbose.
Made the version show up as "devel" during development.

Dodged a nasty interaction between the version substitution marker
and the expansion code by changing the character the expansion code uses
to mark unknown tags to a nul. This is not a bugfix. This is a
hack adjustment. The difference is, one should be proud of a
bugfix. :)

Adjusted the auto-debug code to work on my new development
machine, small.nella.org.

Grapher handles NaN's in the summary without warnings now.